test(ui): split attention presentation and read-marker cases under line budgets
